refactor(domain): 패키지 구조 및 이름 변경 &  bridge_game 추가
make_bridge 패키지 -> bridge_maker로 변경 및 bridge_game 패키지 추가하였습니다.
