+++ src/main/java/bridge/controller/BridgeGameController.java
@@ -3,16 +3,16 @@ package bridge.controller;
 import static bridge.domain.constants.GameCommands.NOTHING;
 import static bridge.domain.constants.MoveResultsSign.MOVE_FAIL;
 
-import bridge.domain.Bridge;
-import bridge.domain.BridgeGame;
+import bridge.domain.bridge_game.Bridge;
+import bridge.domain.bridge_game.BridgeGame;
 import bridge.domain.validation.BridgeSize;
-import bridge.domain.NumberOfChallenges;
-import bridge.domain.Player;
-import bridge.domain.Position;
+import bridge.domain.bridge_game.NumberOfChallenges;
+import bridge.domain.bridge_game.Player;
+import bridge.domain.bridge_game.Position;
 import bridge.domain.constants.GameCommands;
 import bridge.domain.constants.MoveCommands;
-import bridge.domain.make_bridge.BridgeMaker;
-import bridge.domain.MoveResults;
+import bridge.domain.bridge_maker.BridgeMaker;
+import bridge.domain.bridge_game.MoveResults;
 import bridge.view.InputView;
 import bridge.view.OutputView;
