+++ src/main/java/bridge/domain/bridge_game/NumberOfChallenges.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.bridge_game;
 
 public class NumberOfChallenges {
