+++ src/main/java/bridge/domain/bridge_maker/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge.domain.make_bridge;
+package bridge.domain.bridge_maker;
 
 // TODO BridgeRandomNumberGenerator, BridgeNumberGenerator 클래스의 코드는 변경할 수 없다.
 @FunctionalInterface
