+++ src/test/java/bridge/domain/bridge_game/MoveResultsTest.java
@@ -1,10 +1,10 @@
-package bridge.domain;
+package bridge.domain.bridge_game;
 
 import static bridge.domain.constants.MoveCommands.MOVE_DOWN_COMMAND;
 import static bridge.domain.constants.MoveCommands.MOVE_UP_COMMAND;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import bridge.domain.MoveResults;
+import bridge.domain.bridge_game.MoveResults;
 import bridge.domain.constants.MoveCommands;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
