+++ src/test/java/bridge/domain/bridge_game/PositionTest.java
@@ -1,8 +1,9 @@
-package bridge.domain;
+package bridge.domain.bridge_game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import bridge.domain.bridge_game.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
